Removing casting warnings.
This is a partial commit of fixes for the casting warnings.   This set of
changes adds explicit commits where the integer types were not matching.

+++ src/protocols/ab/cip.c
@@ -206,7 +206,7 @@ int cip_encode_path(const char *path, int needs_connection, int plc_type, uint8_
                     return PLCTAG_ERR_BAD_PARAM;
                 }
 
-                *data = tmp;
+                *data = (uint8_t)tmp;
 
                 /*printf("convert_links() link(%d)=%s (%d)\n",i,*links,tmp);*/
 
@@ -243,7 +243,7 @@ int cip_encode_path(const char *path, int needs_connection, int plc_type, uint8_
         data++;
         *data = 0x24;
         data++;
-        *data = dhp_channel;
+        *data = (uint8_t)dhp_channel;
         data++;
         *data = 0x2C;
         data++;
@@ -251,7 +251,7 @@ int cip_encode_path(const char *path, int needs_connection, int plc_type, uint8_
         data++;
         ioi_size += 6;
 
-        *dhp_dest = dest_addr;
+        *dhp_dest = (uint16_t)dest_addr;
     } else if(!has_dhp) {
         if(needs_connection) {
             /*
@@ -302,76 +302,12 @@ int cip_encode_path(const char *path, int needs_connection, int plc_type, uint8_
 
     mem_copy(*conn_path, &tmp_path[0], ioi_size);
 
-    *conn_path_size = ioi_size;
+    *conn_path_size = (uint8_t)ioi_size;
 
     pdebug(DEBUG_INFO, "Done.");
 
     return PLCTAG_STATUS_OK;
 }
-
-
-
-
-
-
-
-//~ char *cip_decode_status(int status)
-//~ {
-    //~ switch(status) {
-        //~ case 0x01:
-            //~ /* FIXME - this should handle the extended error codes */
-            //~ return "Connection error!";
-            //~ break;
-
-        //~ case 0x02:
-            //~ return "Insufficient resources!";
-            //~ break;
-
-        //~ case 0x03:
-            //~ return "Value invalid!"; /* huh? */
-            //~ break;
-
-        //~ case 0x04:
-            //~ return "Bad or indecipherable IOI or tag not found!";
-            //~ break;
-
-        //~ case 0x05:
-            //~ return "Unknown destination, tag or item!";
-            //~ break;
-
-        //~ case 0x06:
-            //~ return "Response too large, partial data transfered!";
-            //~ break;
-
-        //~ case 0x0A:
-            //~ return "Error processing attributes!";
-            //~ break;
-
-        //~ case 0x13:
-            //~ return "Insufficient data/params to process request!";
-            //~ break;
-
-        //~ case 0x1C:
-            //~ return "Insufficient attributes to process request!";
-            //~ break;
-
-        //~ case 0x26:
-            //~ return "IOI word length does not match IOI length processed!";
-            //~ break;
-
-        //~ case 0xFF:
-
-            //~ /* extended status */
-
-        //~ default:
-            //~ return "Unknown error status.";
-            //~ break;
-    //~ }
-
-    //~ return "Unknown error status.";
-//~ }
-
-
 
 
 
@@ -447,7 +383,7 @@ int cip_encode_tag_name(ab_tag_p tag,const char *name)
                 dp++;
 
                 while(isalnum(*p) || *p == '_' || *p == ':') {
-                    *dp = *p;
+                    *dp = (uint8_t)*p;
                     dp++;
                     p++;
                     (*name_len)++;
@@ -488,11 +424,11 @@ int cip_encode_tag_name(ab_tag_p tag,const char *name)
                         /* copy the value in little-endian order */
                         *dp = val & 0xFF;
                         dp++;
-                        *dp = (val >> 8) & 0xFF;
+                        *dp = (uint8_t)((val >> 8) & 0xFF);
                         dp++;
-                        *dp = (val >> 16) & 0xFF;
+                        *dp = (uint8_t)((val >> 16) & 0xFF);
                         dp++;
-                        *dp = (val >> 24) & 0xFF;
+                        *dp = (uint8_t)((val >> 24) & 0xFF);
                         dp++;
                     } else if(val > 0xFF) {
                         *dp = 0x29;
@@ -503,12 +439,12 @@ int cip_encode_tag_name(ab_tag_p tag,const char *name)
                         /* copy the value in little-endian order */
                         *dp = val & 0xFF;
                         dp++;
-                        *dp = (val >> 8) & 0xFF;
+                        *dp = (uint8_t)((val >> 8) & 0xFF);
                         dp++;
                     } else {
                         *dp = 0x28;
                         dp++;  /* 1-byte value */
-                        *dp = val;
+                        *dp = (uint8_t)val;
                         dp++;     /* value */
                     }
